+++ evm-ds/build.rs
@@ -1,10 +1,25 @@
 extern crate protoc_rust;
+use std::path::Path;
 
 fn main() {
+
+    let mut scilla_message_proto = "../src/libPersistence/ScillaMessage.proto";
+    let mut evm_proto = "../src/libUtils/Evm.proto";
+    let mut scilla_message_proto_inc = "../src/libPersistence/";
+    let mut evm_proto_inc = "../src/libUtils/";
+
+    // If we have the files local to this dir, go ahead
+    if Path::new("ScillaMessage.proto").is_file() && Path::new("Evm.proto").is_file() {
+        scilla_message_proto = "ScillaMessage.proto";
+        evm_proto = "Evm.proto";
+        scilla_message_proto_inc = "./";
+        evm_proto_inc = "./";
+    }
+
     protoc_rust::Codegen::new()
         .out_dir("src/protos")
-        .inputs(["../src/libPersistence/ScillaMessage.proto"])
-        .include("../src/libPersistence")
+        .inputs([scilla_message_proto.to_string()])
+        .include(scilla_message_proto_inc.to_string())
         .customize(protoc_rust::Customize {
             carllerche_bytes_for_bytes: Some(true),
             carllerche_bytes_for_string: Some(true),
@@ -15,8 +30,8 @@ fn main() {
 
     protoc_rust::Codegen::new()
         .out_dir("src/protos")
-        .inputs(["../src/libUtils/Evm.proto"])
-        .include("../src/libUtils")
+        .inputs([evm_proto.to_string()])
+        .include(evm_proto_inc.to_string())
         .customize(protoc_rust::Customize {
             carllerche_bytes_for_bytes: Some(true),
             carllerche_bytes_for_string: Some(true),
@@ -25,6 +40,6 @@ fn main() {
         .run()
         .expect("Running protoc failed for EVM.proto");
 
-    println!("cargo:rerun-if-changed=../src/libPersistence/ScillaMessage.proto");
-    println!("cargo:rerun-if-changed=../src/libUtils/Evm.proto");
+    println!("cargo:rerun-if-changed={}", scilla_message_proto);
+    println!("cargo:rerun-if-changed={}", evm_proto);
 }
